reorganize node utils
