+++ src/components/deprecated/Sidebar/PipelineLibrary/GraphComponentLink.tsx
@@ -9,9 +9,9 @@
 import { useStore } from "@xyflow/react";
 import type { CSSProperties, Ref } from "react";
 
-import { updateComponentSpecFromNodes } from "@/components/shared/ReactFlow/FlowGraph/utils/updateComponentSpecFromNodes";
 import type { ComponentSpec } from "@/utils/componentSpec";
 import { componentSpecToYaml } from "@/utils/componentStore";
+import { updateComponentSpecFromNodes } from "@/utils/nodes/updateComponentSpecFromNodes";
 
 interface GraphComponentLinkProps {
   componentSpec: ComponentSpec;
