+++ src/components/shared/ReactFlow/FlowSidebar/sections/SettingsAndActions.tsx
@@ -25,7 +25,7 @@ import { EDITOR_PATH } from "@/routes/router";
 import { useSavePipeline } from "@/services/pipelineService";
 import { componentSpecToYaml } from "@/utils/componentStore";
 
-import { updateComponentSpecFromNodes } from "../../FlowGraph/utils/updateComponentSpecFromNodes";
+import { updateComponentSpecFromNodes } from "../../../../../utils/nodes/updateComponentSpecFromNodes";
 import ImportComponent from "../components/ImportComponent";
 
 const SettingsAndActions = () => {
