+++ src/hooks/useComponentSpecToNodes.ts
@@ -1,8 +1,8 @@
 import { type Node, type NodeChange, useNodesState } from "@xyflow/react";
 import { useEffect } from "react";
 
-import { extractPositionFromAnnotations } from "@/components/shared/ReactFlow/FlowGraph/utils/extractPositionFromAnnotations";
-import type { ComponentTaskNodeCallbacks } from "@/components/shared/ReactFlow/FlowGraph/utils/isComponentTaskNode";
+import type { ComponentTaskNodeCallbacks } from "@/types/taskNode";
+import { extractPositionFromAnnotations } from "@/utils/nodes/extractPositionFromAnnotations";
 
 import type { ComponentSpec, GraphSpec } from "../utils/componentSpec";
