+++ src/types/taskNode.ts
@@ -1,0 +1,6 @@
+import type { ArgumentType } from "@/utils/componentSpec";
+
+export interface ComponentTaskNodeCallbacks {
+  setArguments?: (args: Record<string, ArgumentType>) => void;
+  onDelete: () => void;
+}
