+++ src/utils/nodes/extractPositionFromAnnotations.ts
