+++ src/utils/nodes/updateComponentSpecFromNodes.test.ts
