+++ src/utils/nodes/updateComponentSpecFromNodes.ts
